separerade favorit knappen till en egen komponent
