+++ app/mylist/page.tsx
@@ -1,0 +1,3 @@
+export default function MyList() {
+  return <></>;
+}
